Add separate component for calculating changes
